+++ src/main/java/com/ing/data/cassandra/jdbc/commands/package-info.java
@@ -1,0 +1,29 @@
+/*
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+/**
+ * This package contains the classes used to handle
+ * <a href="https://cassandra.apache.org/doc/stable/cassandra/tools/cqlsh.html#special-commands">special CQL
+ * commands</a>.
+ * <p>
+ *     Supported commands:
+ *     <ul>
+ *         <li>{@code CONSISTENCY [level]}</li>
+ *         <li>{@code SERIAL CONSISTENCY [level]}</li>
+ *         <li>{@code SOURCE <filename>}</li>
+ *         </ul>
+ * </p>
+ */
+package com.ing.data.cassandra.jdbc.commands;