+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add support for the special CQL command `SOURCE <filename>` in `CassandraStatement`.
+- Add a method `CassandraConnection.setOptionSet(OptionSet)` to programmatically define a custom compliance mode option
+  set on a pre-existing connection.
+
 ## [4.14.0] - 2024-12-24
 ### Added
 - Add support for IPv6 addresses in JDBC URL (see PR [#62](https://github.com/ing-bank/cassandra-jdbc-wrapper/pull/62)).